moving new comments + builders into main flow, and starting to update tests

+++ core-domain/src/main/java/uk/ac/ebi/uniprot/domain/uniprot/comment/MassSpectrometryComment.java
@@ -6,9 +6,9 @@ import java.util.List;
 
 public interface MassSpectrometryComment extends Comment {
 
-	public Float getMolWeight();
+    public Double getMolWeight();
 
-	public Float getMolWeightError();
+    public Double getMolWeightError();
 
     public String getNote();
